Added path walking to JsonNavigator. Map persister allows deletion of maps. File deletion works OK if file doesn't exist in the first place. IOMDriver now integrated into IS device framework.

+++ interactivespaces-core/src/main/java/interactivespaces/util/io/Files.java
@@ -275,15 +275,20 @@ public class Files {
 	 * <p>
 	 * If a directory, will recursively delete the directory and its contents.
 	 * 
+	 * <p>
+	 * If the file doesn't exist, nothing happens.
+	 * 
 	 * @param file
-	 *            The file to be deleted.
+	 *            the file to be deleted
 	 */
 	static public final void delete(File file) {
-		if (file.isDirectory()) {
-			deleteDirectoryContents(file);
-		}
+		if (file.exists()) {
+			if (file.isDirectory()) {
+				deleteDirectoryContents(file);
+			}
 
-		file.delete();
+			file.delete();
+		}
 	}
 
 	/**
